rm logging.rs and move its only function to util.rs
