+++ src-tauri/src/commands/util.rs
@@ -87,3 +87,15 @@ pub async fn is_macos_version_15_or_above() -> bool {
     .map(|major| major >= 15)
     .unwrap_or(false)
 }
+
+#[tauri::command]
+pub async fn frontend_log(level: String, log: String) -> Result<(), ()> {
+  match level.as_str() {
+    "debug" => log::debug!("{}", log),
+    "info" => log::info!("{}", log),
+    "warn" => log::warn!("{}", log),
+    "error" => log::error!("{}", log),
+    _ => log::info!("{}", log),
+  }
+  Ok(())
+}
