+++ src-tauri/src/main.rs
@@ -211,8 +211,8 @@ fn main() {
       commands::game::get_furthest_game_milestone,
       commands::game::reset_game_settings,
       commands::game::uninstall_game,
-      commands::logging::frontend_log,
       commands::support::generate_support_package,
+      commands::util::frontend_log,
       commands::util::is_diskspace_requirement_met,
       commands::util::is_macos_version_15_or_above,
       commands::util::is_minimum_vcc_runtime_installed,
